Move NodeAttributes to noderepository
