+++ node-admin/src/main/java/com/yahoo/vespa/hosted/node/admin/configserver/noderepository/NodeAttributes.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.node.admin.nodeagent;
+package com.yahoo.vespa.hosted.node.admin.configserver.noderepository;
 
 import com.yahoo.vespa.hosted.dockerapi.DockerImage;
 
@@ -10,6 +10,7 @@ import java.util.Objects;
 // but using the binding class here would be a layer violation, and would also tie this logic to
 // serialization-related dependencies it needs not have.
 public class NodeAttributes {
+
     private Long restartGeneration = null;
     private Long rebootGeneration = null;
     private DockerImage dockerImage = null;
@@ -42,6 +43,7 @@ public class NodeAttributes {
         this.hardwareDivergence = hardwareDivergence;
         return this;
     }
+
 
     public Long getRestartGeneration() {
         return restartGeneration;
